add better merge and double merge
Add better merge and double merge to decide and process merge a to b or
b to a with less ops expected. Better merge also check and proceed if sa
merge is possible because sa ops is generally less than merge a to b.

Double merge only consider top_nodes and the next_nodes by comparing the
number of ops for ab then ba vs ba then ab.

The total number of ops is reduced to 5800-600ish. Need to implement
more decision function to improve output.