add gpt divide and gpt merge
three way merge checks possiblity of ra, pbrb a to b merge and do the
shortest operations. it divide top subsequence into first and second,
second is the last part of top subsequence that can be merged into
bottom subsequence only by ra, first is the remaining and do pbrb or a
to b merge up to a_top_sub_top vs b_bottom values.

gpt divide and merge classify and optimize merge when n_subsequence is
the greatest power of two. it reconstruct from 0,1 pair to final pairs.

This reduce avg ops to 800 when 100 input 5700 when 500 input. Need to
optimize more