remake for three way merge up to eval
remake whole divide_subs merge_all parts to implement three way merge
which compare three input and create new monotone sequence in the last
connection.

subsequence ind changed to have unique absolute value as id and sign to
represent increasing or decreasing. after assign natural runs,
divide_sub function initally calculate desired sequence of monotone
subsequences and send the exact number and inc/dec as seed. then it
merges remaining subsequences in a to seed and realign.

after divide sub, it is clear to just send seed and merge to toss subs a
to b and b to a until it reaches 1 sub in a.